Prep for next vitepress release.
