+++ docs/.vitepress/theme/index.ts
@@ -1,6 +1,5 @@
 import DefaultTheme from 'vitepress/theme';
 
-import NotFound from './layouts/NotFound.vue';
 import Layout from './layouts/Layout.vue';
 
 import DownloadList from './components/DownloadList.vue';
@@ -9,15 +8,12 @@ import DownloadSorter from './components/DownloadSorter.vue';
 import ExampleField from './components/ExampleField.vue';
 import Home from './components/Home.vue';
 
-import { onMounted } from 'vue';
 import { createPinia } from 'pinia';
-import { useStore } from './store';
 
 import './styles/theme.scss';
 
 export default {
   ...DefaultTheme,
-  NotFound,
   Layout,
 
   enhanceApp(ctx: any): void {
@@ -32,18 +28,5 @@ export default {
     ctx.app.component('Home', Home);
 
     ctx.app.use(createPinia());
-  },
-
-  setup(): void {
-    const store = useStore();
-    const storeValues: string[] = ['Meta', 'EnumValues', 'SetList', 'DeckList']
-
-    onMounted(async (): Promise<void> => {
-      storeValues.forEach(async ( value: string ): Promise<void> => {
-        if ( Object.keys(store[value] ).length === 0) {
-          await store.fetchJSON(value);
-        }
-      });
-    });
-  },
+  }
 };
